requirements.txt
gevent<=24.2.1
ipykernel<=6.29.5
matplotlib<=3.9.2
openai<=1.43.0
psycopg2<=2.9.9
python-dotenv<=1.0.1
redshift-connector<=2.1.3